ref(commands): step 7 | move JobOutput to Core
